Squashed On-The-Go mode (1/2)
On-The-Go turns on the camera of your phone and streams it to an
overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the
content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Change-Id: I3efccb5fac41a946f6cb4992e9181422980ee5e4

Signed-off-by: Chris <sirhc@cypheros.co>