DexOpt: MaterialOpt Dialog
This is a very simple, yet sleek and material fullscreen dialog for the
dexopt portion of initial bootup. It displays your logo image in the
center of the screen, with respective app information at the bottom
alongside it's icon. It also uses the native progress bar for
optimization tracking.

This is a port from Nameless Rom and BrokenOS. Cherrypicking breaks too
many existing changes, so my best bet was to merge it manually and
credit the original authors.

The commit used: 84a08dcc38ccfa9367108f2effbe73e1e1b744ce was reference
to my initial bring up but it had to be modified to work with our base.

Overview:

- Pass application info, current and total number of packages to the
boot message interface
- Turn the dialog into a full screen dialog with custom view

- Awesome logo in the middle, overlay able (in case we need to rebrand)
- core/res/res/drawable/dexopt_brand_logo.xml
- Display app icon, dexopt message and package name of the processed
package
- Show a progressbar on the bottom, showing the current progress
- Disabled app icon loading for low ram devices by default, unless high
end gfx is forced or "persist.sys.dexopt.no_icon"
is set to false

Change-Id: Id07b21b459f8508cfe688139c2afbe27dc4e243a
Signed-off-by: Alexander Martinz
<eviscerationls@gmail.com>
Signed-off-by: Chris<sirhc@cypheros.co>